refactor: LottoInvestment를 application 패키지로 분리하여 domain과 view가 서로를 참조하지 않도록 함
