+++ src/test/java/lotto/application/LottoInvestmentTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.application;
 
 import static org.assertj.core.api.Assertions.*;
